# docs/swagger.yml
openapi: 3.0.3
info:
  title: Battleship API Docs
  description: |-
    This is the API documentation for the Multiplayer Battleship game.
    [kendlbat/battleship-node](https://github.com/kendlbat/battleship-node)
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  version: "1.0"
servers:
  - url: http://localhost:8080/api
...
paths:
...
  /game/create:
    get:
      tags:
        - game
      summary: Creates a new game
      description: Can not be ran while in a valid game
      operationId: createGame
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    description: Status of the API
                    example: "ok"
                  gameId:
                    type: string
                    description: ID of the game
                    example: "797236-0"
                    pattern: "^[0-9]{6}-[0-9]$"
                  token:
                    type: string
                    description: Access token of the player
                    example: "74ce5fbf48f79de0"
                    pattern: "^[0-9a-f]{16}$"
              example:
                status: "ok"
                gameId: "797236-0"
                token: "74ce5fbf48f79de0"
        "400":
          description: Already in a game
  /game/join:
    get:
      tags:
        - game
      summary: Joins a game
      operationId: joinGame
      parameters:
        - name: gameId
          in: query
          description: Game id
          required: true
          schema:
            type: string
            pattern: "^[0-9]{6}-[0-9]$"
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    description: Status of the API
                    example: "ok"
                  gameId:
                    type: string
                    description: ID of the game
                    example: "797236-0"
                    pattern: "^[0-9]{6}-[0-9]$"
                  token:
                    type: string
                    description: Access token of the player
                    example: "74ce5fbf48f79de0"
                    pattern: "^[0-9a-f]{16}$"
              example:
                status: "ok"
                gameId: "797236-0"
                token: "74ce5fbf48f79de0"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                    description: Status of the API
                    example: 400
                  message:
                    type: string
                    description: Error message
                    example: "Game not found"
              example:
                status: 400
                error: "Error message"

        "404":
          description: Not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                    description: Status of the API
                    example: 404
                  message:
                    type: string
                    description: Error message
                    example: "Game not found"
              example:
                status: 404
                error: "Error message"
  /game/status:
    get:
      tags:
        - game
      summary: Get current game status
      operationId: getGameStatus
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    description: Status of the API
                    example: "ok"
                  gameId:
                    type: string
                    description: ID of the game
                    example: "797236-0"
                    pattern: "^[0-9]{6}-[0-9]$"
                  turn:
                    type: number
                    description: Number of the player whose turn it is - 0 if game is not running
                    example: 1
                    enum:
                      - 0
                      - 1
                      - 2
                  state:
                    type: string
                    description: State of the game
                    enum:
                      - waiting
                      - starting
                      - playing
                      - finished
                  player:
                    type: number
                    description: Number of the player
                    example: 1
                    enum:
                      - 1
                      - 2
              examples:
                waiting:
                  summary: Game is waiting for players to join
                  value:
                    status: "ok"
                    state: "waiting"
                    turn: 0
                    gameId: "797236-0"
                    player: 1
                starting:
                  summary: Game is waiting for players to place their ships
                  value:
                    status: "ok"
                    turn: 0
                    state: "starting"
                    gameId: "797236-0"
                    player: 1
                playing:
                  summary: Game is running
                  value:
                    status: "ok"
                    turn: 1
                    state: "playing"
                    gameId: "797236-0"
                    player: 1
                finished:
                  summary: Game is over
                  value:
                    status: "ok"
                    turn: 0
                    state: "finished"
                    gameId: "797236-0"
                    player: 1
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                    description: Status of the API
                    example: 400
                  message:
                    type: string
                    description: Error message
                    example: "Game not found"
              example:
                status: 400
                error: "Error message"
        "403":
          description: Not authorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                    description: Status of the API
                    example: 403
                  message:
                    type: string
                    description: Error message
                    example: "Not authorized"
              example:
                status: 403
                error: "Invalid token"
        "404":
          description: Game not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                    description: Status of the API
                    example: 404
                  message:
                    type: string
                    description: Error message
                    example: "Game not found"
              example:
                status: 404
                error: "Game not found"
      security:
        - gameId: []
        - playerToken: []
  /game/getAvailableShips:
    get:
      tags:
        - game
      summary: Gets an array of unplaced ships
      operationId: getAvailableShips
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    description: Status of the API
                    example: "ok"
                  ships:
                    type: array
                    description: Array of unplaced ships
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          description: ID of the ship
                          example: "carrier"
                        length:
                          type: number
                          description: Length of the ship
                          example: 5
                          minimum: 1
              example:
                status: "ok"
                ships:
                  - id: "carrier"
                    size: 5
                  - id: "battleship"
                    size: 4
                  - id: "cruiser"
                    size: 3
                  - id: "submarine"
                    size: 3
                  - id: "destroyer"
                    size: 2
        "403":
          description: Not authorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                    description: Status of the API
                    example: 403
                  message:
                    type: string
                    description: Error message
                    example: "Invalid token"
              example:
                status: 403
                error: "Invalid token"
        "404":
          description: Game not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                    description: Status of the API
                    example: 404
                  message:
                    type: string
                    description: Error message
                    example: "Game not found"
              example:
                status: 404
                error: "Game not found"
      security:
        - gameId: []
        - playerToken: []
  /game/placeShips:
    post:
      tags:
        - game
      summary: Place ships on the board
      description: The boards coordinates start at 0, 0<br>Normally, board is 8x8<br>Trying to send invalid ships will result in errors
      operationId: placeShips
      requestBody:
        description: Test
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/Ship"
            example:
              ships:
                - x: 0
                  y: 0
                  orientation: "horizontal"
                  id: "carrier"
                  size: 5
                - x: 0
                  y: 3
                  orientation: "vertical"
                  id: "battleship"
                  size: 4
                - x: 3
                  y: 1
                  orientation: "horizontal"
                  id: "cruiser"
                  size: 3
                - x: 4
                  y: 6
                  orientation: "horizontal"
                  id: "submarine"
                  size: 3
                - x: 6
                  y: 1
                  orientation: "vertical"
                  id: "destroyer"
                  size: 2
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    description: Status of the API
                    example: "ok"
                  board:
                    $ref: "#/components/schemas/Board"
              example:
                status: "ok"
                board:
                  - [
                      true,
                      false,
                      false,
                      true,
                      true,
                      true,
                      true,
                      false,
                    ]
                  - [
                      true,
                      false,
                      false,
                      false,
                      false,
                      false,
                      false,
                      false,
                    ]
                  - [
                      true,
                      false,
                      false,
                      false,
                      false,
                      false,
                      false,
                      false,
                    ]
                  - [
                      true,
                      true,
                      false,
                      false,
                      false,
                      false,
                      false,
                      false,
                    ]
                  - [
                      true,
                      true,
                      false,
                      false,
                      false,
                      false,
                      true,
                      false,
                    ]
                  - [
                      false,
                      true,
                      false,
                      false,
                      false,
                      false,
                      true,
                      false,
                    ]
                  - [
                      false,
                      true,
                      true,
                      false,
                      false,
                      false,
                      true,
                      false,
                    ]
                  - [
                      false,
                      false,
                      false,
                      false,
                      false,
                      false,
                      false,
                      false,
                    ]
                  - [
                      false,
                      false,
                      false,
                      false,
                      false,
                      false,
                      false,
                      false,
                    ]
                  - [
                      false,
                      false,
                      false,
                      false,
                      false,
                      false,
                      false,
                      false,
                    ]
        "400":
          description: Malfomed request
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                    description: Status of the API
                    example: 400
                  message:
                    type: string
                    description: Error message
                    example: "Invalid ship"
              example:
                status: 400
                error: "Malformed request"
        "403":
          description: Not authorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                    description: Status of the API
                    example: 403
                  message:
                    type: string
                    description: Error message
                    example: "Invalid token"
              example:
                status: 403
                error: "Invalid token"
        "404":
          description: Game not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                    description: Status of the API
                    example: 404
                  message:
                    type: string
                    description: Error message
                    example: "Game not found"
              example:
                status: 404
                error: "Game not found"
      security:
        - gameId: []
        - playerToken: []
...
components:
  securitySchemes:
    gameId:
      type: apiKey
      in: cookie
      name: gameId
    playerToken:
      type: apiKey
      in: cookie
      name: token
  schemas:
    Ship:
      type: object
      properties:
        id:
          type: string
        x:
          type: integer
        y:
          type: integer
        orientation:
          type: string
          enum:
            - horizontal
            - vertical
        size:
          type: integer
    Board:
      type: array
      items:
        type: array
        items:
          type: boolean
...